requirements.txt
xgboost==1.4.0
PyMySQL==1.0.2
Flask==1.1.2
py2neo==2021.0.1
selenium==3.141.0
numpy==1.19.2
opencv_python==4.5.1.48
pandas==1.1.3
lightgbm==3.2.0
Flask_Cors==3.0.10
matplotlib==3.3.2
torch==1.8.0
torchvision==0.8.2
dgl_cu101==0.6.0.post1
DBUtils==2.0.1
dgl==0.6.1
Pillow==8.3.1
scikit_learn==0.24.2
torch==1.9.0
